Public API

DataTypesBasic.DataTypesBasic

DataTypesBasic.ContextManager

DataTypesBasic.MultipleExceptions

DataTypesBasic.Thrown

DataTypesBasic.@either

DataTypesBasic.DataTypesBasic — Module

NOTE: in julia we can represent closed abstract types with two different systems:

either Union{Type1, Type2}

or with a standard abstract type, where all type definitions are on the abstract type level

(if you would only define functions on the concretetypes, the the inference engine will do weird things as it assumes there might be another concretetype in the future)

Approach 1.

Advantages: Providing definitions for the concrete is actually enough for typeinference. More code reuse. Disadvantages: The typeinference like for Dict(:a => Some(1), :b => None()) would infer Dict{Symbol, Any} instead of Dict{Symbol, Union{Some, None}}.

The internal details for this is that in many cases Base.promote_typejoin(type1, type2) is used to come up with a common type. While the type hierarchy is used for this, Base.promote_typejoin of two unrelated types will always result in Any.

You would need to overload promote_typejoin to support Dict(:a => Identity(1), :b => nothing) isa Dict{Symbol, Option{Int}}.

Approach 2.

Advantages: Dict(:a => Some(1), :b => None()) would indeed infer Dict{Symbol, Option} Disadvantages: you need to be careful to always implement functionalities first on separate functions unique to the sealed type, and then point generic functions to the specific one via genericfunction(o::Option) = optionfunction(o)

As we managed to make the promote_typejoin work as intended for Union, we follow Approach 1, in order to enable way higher code-reuse and datatype-reuse.

DataTypesBasic.ContextManager — Type

function which expects one argument, which itself is a function. Think of it like the following:

function contextmanagerready(cont)
  # ... do something before
  value = ... # create some value to work on later
  result = cont(value)  # pass the value to the continuation function (think like `yield`)
  # ... do something before exiting, e.g. cleaning up
  result # IMPORTANT: always return the result of the `cont` function
end

Now you can wrap it into ContextManager(contextmanagerready) and you can use all the context manager functionalities right away

DataTypesBasic.MultipleExceptions — Type

Combine Multiple Exceptions

DataTypesBasic.Thrown — Type

Thrown is like Exception, however can also cary stacktraces

DataTypesBasic.@either — Macro

@either true ? "right" : Symbol("left")
@either if false
  "right"
else
  :left
end

Simple macro to reuse ? operator and simple if-else for constructing Either.